ADR-830 fixing page content, validations  and controller

+++ app/pages/DeclareIrishWhiskyPage.scala
@@ -18,7 +18,7 @@ package pages
 
 import play.api.libs.json.JsPath
 
-case object DeclareIrishWhiskyPage extends QuestionPage[Int] {
+case object DeclareIrishWhiskyPage extends QuestionPage[BigDecimal] {
 
   override def path: JsPath = JsPath \ toString
